+++ test/data/src/Wierd.java
@@ -6,8 +6,15 @@ public class Wierd {
   public static C. D . @Ann E <?> x;
   public static F. G . @Ann H <?> y;
 
-  public <T> C.D.E m () { return null;}
-  public <T> F.G.H n () { return null;}
+  public <T> C.D.@Ann E m () { return null;}
+  public <T> F.G.@Ann H n () { return null;}
+
+  public class X <Q extends C.D.@Ann E<?>> extends @Ann Object {
+  }
+  
+  public class Y <Q extends F.G.@Ann H<?>>{
+  }
+  
   
   public class C {
     public class D {
@@ -18,7 +25,7 @@ public class Wierd {
   
   public static class F {
     public static class G {
-      public class H <T> {
+      public interface H <T> {
       }
     }
   }
